put the mainscreen in a seperate file
